requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile -o requirements.txt requirements-base.in requirements.in
aiofiles==23.2.1
    # via gcloud-aio-storage
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via gcloud-aio-auth
aiosignal==1.4.0
    # via aiohttp
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   starlette
async-timeout==5.0.1
    # via
    #   aiohttp
    #   redis
attrs==25.4.0
    # via aiohttp
backoff==2.2.1
    # via gcloud-aio-auth
cachetools==6.2.1
    # via google-auth
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
cffi==2.0.0
    # via cryptography
chardet==5.2.0
    # via gcloud-aio-auth
click==8.1.8
    # via
    #   -r requirements-base.in
    #   uvicorn
cryptography==46.0.3
    # via gcloud-aio-auth
environs==9.5.0
    # via
    #   -r requirements-base.in
    #   gundi-client-v2
exceptiongroup==1.3.0
    # via anyio
fastapi==0.103.2
    # via -r requirements-base.in
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
gcloud-aio-auth==5.4.2
    # via
    #   gcloud-aio-pubsub
    #   gcloud-aio-storage
gcloud-aio-pubsub==6.0.1
    # via -r requirements-base.in
gcloud-aio-storage==9.3.0
    # via -r requirements.in
google-auth==2.41.1
    # via -r requirements.in
gundi-client-v2==2.4.1
    # via -r requirements-base.in
gundi-core==1.11.2
    # via
    #   -r requirements-base.in
    #   gundi-client-v2
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   gundi-client-v2
    #   respx
idna==3.11
    # via
    #   anyio
    #   httpx
    #   yarl
marshmallow==3.22.0
    # via
    #   -r requirements-base.in
    #   environs
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
packaging==25.0
    # via marshmallow
prometheus-client==0.23.1
    # via gcloud-aio-pubsub
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via
    #   gcloud-aio-storage
    #   google-auth
pycparser==2.23
    # via cffi
pydantic==1.10.24
    # via
    #   -r requirements-base.in
    #   fastapi
    #   gundi-client-v2
    #   gundi-core
pyjq==2.6.0
    # via -r requirements-base.in
pyjwt==2.10.1
    # via gcloud-aio-auth
python-dotenv==1.1.1
    # via environs
python-json-logger==2.0.7
    # via -r requirements-base.in
redis==5.0.8
    # via -r requirements-base.in
respx==0.22.0
    # via gundi-client-v2
rsa==4.9.1
    # via
    #   gcloud-aio-storage
    #   google-auth
sniffio==1.3.1
    # via anyio
stamina==23.2.0
    # via -r requirements-base.in
starlette==0.27.0
    # via fastapi
tenacity==9.1.2
    # via stamina
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   cryptography
    #   exceptiongroup
    #   fastapi
    #   multidict
    #   pydantic
    #   uvicorn
uvicorn==0.23.2
    # via -r requirements-base.in
yarl==1.22.0
    # via aiohttp
